requirements.txt
streamlit
google-generativeai
instaloader
Pillow